requirements.txt
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.2.0
tqdm==4.62.3
h5py==3.6.0
imageio==2.9.0
patchify==0.2.3
opencv-python==4.6.0.66
PyQt5==5.15.7
ipykernel==6.9.1
seaborn==0.11.2
pandas==1.4.3
scikit-image==0.18.3
